refactor: add Account, Branch and BranchEntry types
